W-23516605 refactor(native-lib): split Node binding into focused modules
Behavior-preserving extraction of src/index.ts to improve readability and
expose pure logic for unit testing. No behavior changes; public export
surface and the existing integration suite are unchanged.

- errors.ts: DataWeaveError / DataWeaveScriptError.
- reader.ts: createChunkReader — the pure sync/async iterable → pull-based
  byte reader used by runTransform.
- stream.ts: streamFromNative — the pure callback→async-generator bridge,
  deduplicating the producer/consumer/backpressure block previously copied
  between runStreaming and runTransform. The native call is injected so the
  bridge is testable without the addon.
- dataweave.ts: the DataWeave class and module-level singleton API; methods
  reduced to wiring over the extracted helpers. libPath is now readonly.
- index.ts: thin barrel re-exporting the same public names (main =
  dist/index.js consumers unaffected).

All modules carry TSDoc. Verified: tsc clean, 35 unit tests pass, 14
integration tests pass against the real native library.
